pyproject.toml
[project]
name = "flowbot3d"
version = "0.1.0"
description = "A Python Package Template"
readme = "README.md"
requires-python = ">=3.6"
license = {file = "LICENSE.txt"}
authors = [
  {email = "baeisner@andrew.cmu.edu", name = "Ben Eisner"}
]
dependencies = [
  "numpy >= 1.22.4",
  "torch >= 1.11.0",
  "pytorch-lightning >= 1.6.4",
  "typer >= 0.4.0",
  "torch-scatter >= 2.0.9",
  "torch-sparse >= 0.6",
  "torch-cluster >= 1.6.0",
  "torch-spline-conv >= 1.2.1",
  "torch-geometric >= 2.0",
  "rpad-partnet-mobility-utils[pybullet] @ git+https://github.com/r-pad/partnet_mobility_utils",
  "rpad-visualize-3d @ git+https://github.com/r-pad/visualize_3d",
  "rpad-pyg @ git+https://github.com/r-pad/pyg_libs",
]

[build-system]
requires = [
  "setuptools >= 62.3.2",
  "setuptools-scm",
  "wheel",
]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
develop = [
  "pytest",
  "black >= 22.3.0",
  "isort",
  "mypy",
  "pylint",
  "autoflake",
  "pre-commit",
  "jupyterlab",
]
build_docs = [
  "mkdocs-material",
  "mkdocstrings[python]",
]

# This is required to allow us to have notebooks/ at the top level.
[tool.setuptools]
packages = ["flowbot3d"]

[tool.isort]
profile = "black"

[tool.mypy]
python_version = 3.8
warn_return_any = true
warn_unused_configs = true

# # Uncomment this when you have imports for mypy to ignore.
# [[tool.mypy.overrides]]
# module = [
# ]
# ignore_missing_imports = true
